// frontend/app/src/app/models/korisnik.ts
export class Korisnik {
    korisnicko_ime: string = "";
    lozinka: string = "";
    ime: string = "";
    prezime: string = "";
    pol: string = "";
    adresa: string = "";
    kontakt_telefon: string = "";
    imejl: string = "";
    profilna_slika: string = "";
    kreditna_kartica: string = "";
    tip: string = "";
    vlasnik_status: string = "";
    dekorater_status: string = "";
    dekorater_firma: string = "";
}